+++ packages/error-codes/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @module-federation/error-codes
 
+## 2.8.2
+
 ## 2.8.1
 
 ## 2.8.0