+++ packages/node/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/node
 
+## 2.7.49
+
+### Patch Changes
+
+- @module-federation/runtime@2.8.2
+- @module-federation/enhanced@2.8.2
+- @module-federation/sdk@2.8.2
+
 ## 2.7.48
 
 ### Patch Changes